+++ main.py
@@ -4,10 +4,10 @@ import serial
 import argparse
 from utilities.log_trim import trim_log_file
 from scanner_adapters.scanner_utils import find_all_scanner_ports
-from scanner_adapters.bc125atAdapter import BC125ATAdapter
-from scanner_adapters.bcd325p2Adapter import BCD325P2Adapter
-from scanner_adapters.sds100Adapter import SDS100Adapter
-from scanner_adapters.aordv1Adapter import AORDV1Adapter
+from scanner_adapters.adapter_bc125at import BC125ATAdapter
+from scanner_adapters.adapter_bcd325p2 import BCD325P2Adapter
+from scanner_adapters.adapter_sds100 import SDS100Adapter
+from scanner_adapters.adapter_ardv1 import AORDV1Adapter
 from command_registry import build_command_table
 from utilities.readlineSetup import initialize_readline
