+++ scanner_adapters/adapter_ardv1.py
